Package moves to separate queries from web
